Refactor property handlers into focused files
