+++ internal/handlers/navigation_properties.go
@@ -319,7 +319,7 @@ func (h *EntityHandler) handleNavigationCollectionItem(w http.ResponseWriter, r 
 		}
 		return
 	}
-	
+
 	// Preload the navigation property to verify the relationship
 	parentDB = parentDB.Preload(navProp.Name)
 	if err := parentDB.First(parent).Error; err != nil {
@@ -341,7 +341,7 @@ func (h *EntityHandler) handleNavigationCollectionItem(w http.ResponseWriter, r 
 	// Check if the target key exists in the collection
 	found := false
 	var targetEntity interface{}
-	
+
 	if navFieldValue.Kind() == reflect.Slice {
 		for i := 0; i < navFieldValue.Len(); i++ {
 			item := navFieldValue.Index(i)
@@ -453,711 +453,13 @@ func (h *EntityHandler) handleOptionsNavigationPropertyCount(w http.ResponseWrit
 	w.WriteHeader(http.StatusOK)
 }
 
-// findNavigationProperty finds a navigation property by name in the entity metadata
-func (h *EntityHandler) findNavigationProperty(navigationProperty string) *metadata.PropertyMetadata {
-	return h.metadata.FindNavigationProperty(navigationProperty)
-}
-
-// findStructuralProperty finds a structural (non-navigation) property by name in the entity metadata
-func (h *EntityHandler) findStructuralProperty(propertyName string) *metadata.PropertyMetadata {
-	return h.metadata.FindStructuralProperty(propertyName)
-}
-
-// IsNavigationProperty checks if a property name is a navigation property
-func (h *EntityHandler) IsNavigationProperty(propertyName string) bool {
-	// Parse out any key from the property name (e.g., "RelatedProducts(2)" -> "RelatedProducts")
-	navPropName, _ := h.parseNavigationPropertyWithKey(propertyName)
-	return h.findNavigationProperty(navPropName) != nil
-}
-
-// IsStructuralProperty checks if a property name is a structural property
-func (h *EntityHandler) IsStructuralProperty(propertyName string) bool {
-	return h.findStructuralProperty(propertyName) != nil
-}
-
-// findComplexTypeProperty finds a complex type property by name in the entity metadata
-func (h *EntityHandler) findComplexTypeProperty(propertyName string) *metadata.PropertyMetadata {
-	return h.metadata.FindComplexTypeProperty(propertyName)
-}
-
-// IsComplexTypeProperty checks if a property name is a complex type property
-func (h *EntityHandler) IsComplexTypeProperty(propertyName string) bool {
-	return h.findComplexTypeProperty(propertyName) != nil
-}
-
-// HandleComplexTypeProperty handles GET, HEAD, and OPTIONS requests for complex type properties (e.g., Products(1)/ShippingAddress)
-// propertySegments represents the navigation path segments without $value/$ref/$count (e.g., ["ShippingAddress", "City"]).
-func (h *EntityHandler) HandleComplexTypeProperty(w http.ResponseWriter, r *http.Request, entityKey string, propertySegments []string, isValue bool) {
-	if len(propertySegments) == 0 {
-		h.writePropertyNotFoundError(w, "")
-		return
-	}
-
-	switch r.Method {
-	case http.MethodGet, http.MethodHead:
-		h.handleGetComplexTypeProperty(w, r, entityKey, propertySegments, isValue)
-	case http.MethodOptions:
-		h.handleOptionsComplexTypeProperty(w)
-	default:
-		h.writeMethodNotAllowedError(w, r.Method, "complex property access")
-	}
-}
-
-// handleGetComplexTypeProperty resolves and writes a complex type property response
-func (h *EntityHandler) handleGetComplexTypeProperty(w http.ResponseWriter, r *http.Request, entityKey string, propertySegments []string, isValue bool) {
-	rootName := propertySegments[0]
-	complexProp := h.findComplexTypeProperty(rootName)
-	if complexProp == nil {
-		h.writePropertyNotFoundError(w, rootName)
-		return
-	}
-
-	if len(propertySegments) == 1 && isValue {
-		if err := response.WriteError(w, http.StatusBadRequest, "Invalid request",
-			"$value is not supported on complex properties"); err != nil {
-			fmt.Printf(LogMsgErrorWritingErrorResponse, err)
-		}
-		return
-	}
-
-	fieldValue, err := h.fetchComplexPropertyValue(w, entityKey, complexProp)
-	if err != nil {
-		return
-	}
-
-	// Handle root null complex property
-	if isNilPointer(fieldValue) {
-		h.writeNoContentForNullComplex(w, r)
-		return
-	}
-
-	// Prepare traversal state
-	contextSegments := []string{complexProp.JsonName}
-	currentValue := dereferenceValue(fieldValue)
-	currentType := dereferenceType(complexProp.Type)
-
-	// Traverse nested segments, if any
-	for idx, segment := range propertySegments[1:] {
-		resolved, resolvedType, resolvedJSONName, ok := resolveStructField(currentValue, segment)
-		if !ok {
-			// Support maps as well as structs
-			if currentValue.Kind() == reflect.Map {
-				resolved, resolvedType, resolvedJSONName, ok = resolveMapField(currentValue, segment)
-			}
-		}
-
-		if !ok {
-			h.writePropertyNotFoundError(w, segment)
-			return
-		}
-
-		contextSegments = append(contextSegments, resolvedJSONName)
-		currentValue = resolved
-		currentType = resolvedType
-
-		// If there are more segments to traverse, ensure we can continue
-		if idx < len(propertySegments[1:])-1 {
-			if isNilPointer(currentValue) {
-				h.writeComplexSegmentNullError(w, contextSegments)
-				return
-			}
-			currentValue = dereferenceValue(currentValue)
-			currentType = dereferenceType(currentType)
-		}
-	}
-
-	h.writeResolvedComplexValue(w, r, entityKey, contextSegments, currentValue, currentType, isValue)
-}
-
-// handleOptionsComplexTypeProperty handles OPTIONS requests for complex properties
-func (h *EntityHandler) handleOptionsComplexTypeProperty(w http.ResponseWriter) {
-	w.Header().Set("Allow", "GET, HEAD, OPTIONS")
-	w.WriteHeader(http.StatusOK)
-}
-
-// HandleStructuralProperty handles GET and OPTIONS requests for structural properties (e.g., Products(1)/Name)
-// When isValue is true, returns the raw property value without JSON wrapper (e.g., Products(1)/Name/$value)
-func (h *EntityHandler) HandleStructuralProperty(w http.ResponseWriter, r *http.Request, entityKey string, propertyName string, isValue bool) {
-	switch r.Method {
-	case http.MethodGet, http.MethodHead:
-		h.handleGetStructuralProperty(w, r, entityKey, propertyName, isValue)
-	case http.MethodOptions:
-		h.handleOptionsStructuralProperty(w)
-	default:
-		h.writeMethodNotAllowedError(w, r.Method, "property access")
-	}
-}
-
-// handleGetStructuralProperty handles GET requests for structural properties
-func (h *EntityHandler) handleGetStructuralProperty(w http.ResponseWriter, r *http.Request, entityKey string, propertyName string, isValue bool) {
-	// Find and validate the structural property
-	prop := h.findStructuralProperty(propertyName)
-	if prop == nil {
-		h.writePropertyNotFoundError(w, propertyName)
-		return
-	}
-
-	// Fetch property value
-	fieldValue, err := h.fetchPropertyValue(w, entityKey, prop)
-	if err != nil {
-		return // Error already written
-	}
-
-	// Write response
-	if isValue {
-		h.writeRawPropertyValue(w, r, prop, fieldValue)
-	} else {
-		h.writePropertyResponse(w, r, entityKey, prop, fieldValue)
-	}
-}
-
-// handleOptionsStructuralProperty handles OPTIONS requests for structural properties
-func (h *EntityHandler) handleOptionsStructuralProperty(w http.ResponseWriter) {
-	w.Header().Set("Allow", "GET, HEAD, OPTIONS")
-	w.WriteHeader(http.StatusOK)
-}
-
-// fetchComplexPropertyValue fetches a complex property value from an entity without applying select clauses
-func (h *EntityHandler) fetchComplexPropertyValue(w http.ResponseWriter, entityKey string, prop *metadata.PropertyMetadata) (reflect.Value, error) {
-	entity := reflect.New(h.metadata.EntityType).Interface()
-	
-	var db *gorm.DB
-	var err error
-	
-	// Handle singleton case where entityKey is empty
-	if h.metadata.IsSingleton && entityKey == "" {
-		// For singletons, we don't use a key query, just fetch the first (and only) record
-		db = h.db
-	} else {
-		// For regular entities, build the key query
-		db, err = h.buildKeyQuery(entityKey)
-		if err != nil {
-			if writeErr := response.WriteError(w, http.StatusBadRequest, ErrMsgInvalidKey, err.Error()); writeErr != nil {
-				fmt.Printf(LogMsgErrorWritingErrorResponse, writeErr)
-			}
-			return reflect.Value{}, err
-		}
-	}
-
-	if err := db.First(entity).Error; err != nil {
-		h.handlePropertyFetchError(w, err, entityKey)
-		return reflect.Value{}, err
-	}
-
-	entityValue := reflect.ValueOf(entity).Elem()
-	fieldValue := entityValue.FieldByName(prop.Name)
-	if !fieldValue.IsValid() {
-		if writeErr := response.WriteError(w, http.StatusInternalServerError, ErrMsgInternalError,
-			"Could not access complex property"); writeErr != nil {
-			fmt.Printf(LogMsgErrorWritingErrorResponse, writeErr)
-		}
-		return reflect.Value{}, fmt.Errorf("invalid field")
-	}
-
-	return fieldValue, nil
-}
-
-// writeNoContentForNullComplex writes a 204 No Content response for null complex properties
-func (h *EntityHandler) writeNoContentForNullComplex(w http.ResponseWriter, r *http.Request) {
-	metadataLevel := response.GetODataMetadataLevel(r)
-	w.Header().Set(HeaderContentType, fmt.Sprintf("application/json;odata.metadata=%s", metadataLevel))
-	w.WriteHeader(http.StatusNoContent)
-}
-
-// writeComplexSegmentNullError writes an error when a nested complex segment is null
-func (h *EntityHandler) writeComplexSegmentNullError(w http.ResponseWriter, contextSegments []string) {
-	path := strings.Join(contextSegments, "/")
-	if err := response.WriteError(w, http.StatusNotFound, "Property not found",
-		fmt.Sprintf("Complex property path '%s' is null", path)); err != nil {
-		fmt.Printf(LogMsgErrorWritingErrorResponse, err)
-	}
-}
-
-// writeResolvedComplexValue writes the resolved complex property (or nested primitive) response
-func (h *EntityHandler) writeResolvedComplexValue(w http.ResponseWriter, r *http.Request, entityKey string, contextSegments []string, value reflect.Value, valueType reflect.Type, isValue bool) {
-	if isComplexValue(value, valueType) {
-		if isValue {
-			if err := response.WriteError(w, http.StatusBadRequest, "Invalid request",
-				"$value is not supported on complex properties"); err != nil {
-				fmt.Printf(LogMsgErrorWritingErrorResponse, err)
-			}
-			return
-		}
-
-		if isNilPointer(value) {
-			h.writeNoContentForNullComplex(w, r)
-			return
-		}
-
-		resolvedValue := dereferenceValue(value)
-		h.writeComplexValueResponse(w, r, entityKey, contextSegments, resolvedValue)
-		return
-	}
-
-	// Primitive value resolution
-	resolvedValue := dereferenceValue(value)
-
-	if isValue {
-		h.writeRawPrimitiveValue(w, r, resolvedValue)
-		return
-	}
-
-	h.writePrimitiveComplexPropertyResponse(w, r, entityKey, contextSegments, value)
-}
-
-// writeComplexValueResponse serializes a complex value (struct or map) with @odata.context
-func (h *EntityHandler) writeComplexValueResponse(w http.ResponseWriter, r *http.Request, entityKey string, contextSegments []string, value reflect.Value) {
-	metadataLevel := response.GetODataMetadataLevel(r)
-	w.Header().Set(HeaderContentType, fmt.Sprintf("application/json;odata.metadata=%s", metadataLevel))
-	w.WriteHeader(http.StatusOK)
-
-	if r.Method == http.MethodHead {
-		return
-	}
-
-	contextPath := strings.Join(contextSegments, "/")
-	responseMap := make(map[string]interface{})
-	if metadataLevel != "none" {
-		contextURL := fmt.Sprintf("%s/$metadata#%s(%s)/%s", response.BuildBaseURL(r), h.metadata.EntitySetName, entityKey, contextPath)
-		responseMap[ODataContextProperty] = contextURL
-	}
-
-	switch value.Kind() {
-	case reflect.Struct:
-		structMap := structValueToMap(value)
-		for k, v := range structMap {
-			responseMap[k] = v
-		}
-	case reflect.Map:
-		iter := value.MapRange()
-		for iter.Next() {
-			key := iter.Key()
-			if key.Kind() == reflect.String {
-				responseMap[key.String()] = iter.Value().Interface()
-			}
-		}
-	default:
-		responseMap["value"] = value.Interface()
-	}
-
-	if err := json.NewEncoder(w).Encode(responseMap); err != nil {
-		fmt.Printf("Error writing complex property response: %v\n", err)
-	}
-}
-
-// writePrimitiveComplexPropertyResponse writes a primitive property (nested within a complex type)
-func (h *EntityHandler) writePrimitiveComplexPropertyResponse(w http.ResponseWriter, r *http.Request, entityKey string, contextSegments []string, value reflect.Value) {
-	metadataLevel := response.GetODataMetadataLevel(r)
-	w.Header().Set(HeaderContentType, fmt.Sprintf("application/json;odata.metadata=%s", metadataLevel))
-	w.WriteHeader(http.StatusOK)
-
-	if r.Method == http.MethodHead {
-		return
-	}
-
-	var valueInterface interface{}
-	if !value.IsValid() {
-		valueInterface = nil
-	} else if value.Kind() == reflect.Ptr {
-		if value.IsNil() {
-			valueInterface = nil
-		} else {
-			valueInterface = value.Elem().Interface()
-		}
-	} else {
-		valueInterface = value.Interface()
-	}
-
-	responseBody := map[string]interface{}{
-		"value": valueInterface,
-	}
-
-	if metadataLevel != "none" {
-		contextURL := fmt.Sprintf("%s/$metadata#%s(%s)/%s", response.BuildBaseURL(r), h.metadata.EntitySetName, entityKey, strings.Join(contextSegments, "/"))
-		responseBody[ODataContextProperty] = contextURL
-	}
-
-	if err := json.NewEncoder(w).Encode(responseBody); err != nil {
-		fmt.Printf("Error writing complex primitive property response: %v\n", err)
-	}
-}
-
-// writeRawPrimitiveValue writes a primitive value in raw form for $value requests
-func (h *EntityHandler) writeRawPrimitiveValue(w http.ResponseWriter, r *http.Request, value reflect.Value) {
-	if !value.IsValid() {
-		w.WriteHeader(http.StatusNoContent)
-		return
-	}
-
-	switch value.Kind() {
-	case reflect.String:
-		w.Header().Set(HeaderContentType, ContentTypePlainText)
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64,
-		reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64,
-		reflect.Float32, reflect.Float64,
-		reflect.Bool:
-		w.Header().Set(HeaderContentType, ContentTypePlainText)
-	default:
-		w.Header().Set(HeaderContentType, "application/octet-stream")
-	}
-
-	w.WriteHeader(http.StatusOK)
-
-	if r.Method == http.MethodHead {
-		return
-	}
-
-	if _, err := fmt.Fprintf(w, "%v", value.Interface()); err != nil {
-		fmt.Printf("Error writing raw primitive value: %v\n", err)
-	}
-}
-
-// resolveStructField resolves a field from a struct by segment name, returning the reflect.Value, type, and canonical JSON name
-func resolveStructField(value reflect.Value, segment string) (reflect.Value, reflect.Type, string, bool) {
-	if value.Kind() != reflect.Struct {
-		return reflect.Value{}, nil, "", false
-	}
-
-	valueType := value.Type()
-	lowerSegment := strings.ToLower(segment)
-	for i := 0; i < value.NumField(); i++ {
-		field := valueType.Field(i)
-		if !field.IsExported() {
-			continue
-		}
-
-		jsonName := getJSONFieldName(field)
-		candidates := []string{field.Name, jsonName}
-		for _, candidate := range candidates {
-			if candidate == "" || candidate == "-" {
-				continue
-			}
-			if candidate == segment || strings.ToLower(candidate) == lowerSegment {
-				return value.Field(i), field.Type, jsonName, true
-			}
-		}
-	}
-
-	return reflect.Value{}, nil, "", false
-}
-
-// resolveMapField resolves a value from a map with string keys
-func resolveMapField(value reflect.Value, segment string) (reflect.Value, reflect.Type, string, bool) {
-	if value.Kind() != reflect.Map || value.Type().Key().Kind() != reflect.String {
-		return reflect.Value{}, nil, "", false
-	}
-
-	mapValue := value.MapIndex(reflect.ValueOf(segment))
-	if mapValue.IsValid() {
-		return mapValue, mapValue.Type(), segment, true
-	}
-
-	lowerSegment := strings.ToLower(segment)
-	iter := value.MapRange()
-	for iter.Next() {
-		key := iter.Key()
-		if strings.ToLower(key.String()) == lowerSegment {
-			val := iter.Value()
-			return val, val.Type(), key.String(), true
-		}
-	}
-
-	return reflect.Value{}, nil, "", false
-}
-
-// isNilPointer checks if a value is a nil pointer or interface
-func isNilPointer(value reflect.Value) bool {
-	if !value.IsValid() {
-		return true
-	}
-
-	switch value.Kind() {
-	case reflect.Ptr, reflect.Interface, reflect.Map, reflect.Slice:
-		return value.IsNil()
-	default:
-		return false
-	}
-}
-
-// dereferenceValue dereferences pointer values until a non-pointer is reached
-func dereferenceValue(value reflect.Value) reflect.Value {
-	for value.IsValid() && value.Kind() == reflect.Ptr {
-		if value.IsNil() {
-			return reflect.Value{}
-		}
-		value = value.Elem()
-	}
-	return value
-}
-
-// dereferenceType dereferences pointer types until a non-pointer is reached
-func dereferenceType(t reflect.Type) reflect.Type {
-	if t == nil {
-		return nil
-	}
-	for t.Kind() == reflect.Ptr {
-		t = t.Elem()
-	}
-	return t
-}
-
-// isComplexValue determines if a resolved value should be treated as a complex object
-func isComplexValue(value reflect.Value, valueType reflect.Type) bool {
-	if !value.IsValid() {
-		if valueType == nil {
-			return false
-		}
-		value = reflect.New(dereferenceType(valueType)).Elem()
-	}
-
-	if value.Kind() == reflect.Map {
-		return true
-	}
-
-	resolvedType := valueType
-	if resolvedType == nil {
-		resolvedType = value.Type()
-	}
-
-	resolvedType = dereferenceType(resolvedType)
-	if resolvedType == nil {
-		return false
-	}
-
-	if resolvedType.Kind() != reflect.Struct {
-		return false
-	}
-
-	// Treat time.Time (and similar) as primitive despite being structs
-	if resolvedType.PkgPath() == "time" && resolvedType.Name() == "Time" {
-		return false
-	}
-
-	return true
-}
-
-// structValueToMap converts a struct to a map using JSON tag names
-func structValueToMap(value reflect.Value) map[string]interface{} {
-	result := make(map[string]interface{})
-	valueType := value.Type()
-	for i := 0; i < value.NumField(); i++ {
-		field := valueType.Field(i)
-		if !field.IsExported() {
-			continue
-		}
-
-		jsonName := getJSONFieldName(field)
-		if jsonName == "-" {
-			continue
-		}
-		if jsonName == "" {
-			jsonName = field.Name
-		}
-
-		result[jsonName] = value.Field(i).Interface()
-	}
-	return result
-}
-
-// getJSONFieldName extracts the JSON field name from a struct field
-func getJSONFieldName(field reflect.StructField) string {
-	tag := field.Tag.Get("json")
-	if tag == "" {
-		return field.Name
-	}
-
-	parts := strings.Split(tag, ",")
-	if len(parts) == 0 || parts[0] == "" {
-		return field.Name
-	}
-
-	return parts[0]
-}
-
-// writeMethodNotAllowedError writes a method not allowed error for a specific context
-func (h *EntityHandler) writeMethodNotAllowedError(w http.ResponseWriter, method, context string) {
-	if err := response.WriteError(w, http.StatusMethodNotAllowed, ErrMsgMethodNotAllowed,
-		fmt.Sprintf("Method %s is not supported for %s", method, context)); err != nil {
-		fmt.Printf(LogMsgErrorWritingErrorResponse, err)
-	}
-}
-
-// writePropertyNotFoundError writes a property not found error
-func (h *EntityHandler) writePropertyNotFoundError(w http.ResponseWriter, propertyName string) {
-	if err := response.WriteError(w, http.StatusNotFound, "Property not found",
-		fmt.Sprintf("'%s' is not a valid property for %s", propertyName, h.metadata.EntitySetName)); err != nil {
-		fmt.Printf(LogMsgErrorWritingErrorResponse, err)
-	}
-}
-
-// fetchPropertyValue fetches a property value from an entity
-func (h *EntityHandler) fetchPropertyValue(w http.ResponseWriter, entityKey string, prop *metadata.PropertyMetadata) (reflect.Value, error) {
-	entity := reflect.New(h.metadata.EntityType).Interface()
-	
-	var db *gorm.DB
-	var err error
-	
-	// Handle singleton case where entityKey is empty
-	if h.metadata.IsSingleton && entityKey == "" {
-		// For singletons, we don't use a key query, just fetch the first (and only) record
-		db = h.db
-	} else {
-		// For regular entities, build the key query
-		db, err = h.buildKeyQuery(entityKey)
-		if err != nil {
-			if writeErr := response.WriteError(w, http.StatusBadRequest, ErrMsgInvalidKey, err.Error()); writeErr != nil {
-				fmt.Printf(LogMsgErrorWritingErrorResponse, writeErr)
-			}
-			return reflect.Value{}, err
-		}
-	}
-
-	db = h.applyStructuralPropertySelect(db, prop)
-
-	if err := db.First(entity).Error; err != nil {
-		h.handlePropertyFetchError(w, err, entityKey)
-		return reflect.Value{}, err
-	}
-
-	entityValue := reflect.ValueOf(entity).Elem()
-	fieldValue := entityValue.FieldByName(prop.Name)
-	if !fieldValue.IsValid() {
-		if err := response.WriteError(w, http.StatusInternalServerError, ErrMsgInternalError,
-			"Could not access property"); err != nil {
-			fmt.Printf(LogMsgErrorWritingErrorResponse, err)
-		}
-		return reflect.Value{}, fmt.Errorf("invalid field")
-	}
-
-	return fieldValue, nil
-}
-
-// handlePropertyFetchError handles errors when fetching a property
-func (h *EntityHandler) handlePropertyFetchError(w http.ResponseWriter, err error, entityKey string) {
-	if err == gorm.ErrRecordNotFound {
-		var errorMessage string
-		if h.metadata.IsSingleton {
-			errorMessage = fmt.Sprintf("Singleton '%s' not found", h.metadata.SingletonName)
-		} else {
-			errorMessage = fmt.Sprintf("Entity with key '%s' not found", entityKey)
-		}
-		
-		if writeErr := response.WriteError(w, http.StatusNotFound, ErrMsgEntityNotFound, errorMessage); writeErr != nil {
-			fmt.Printf(LogMsgErrorWritingErrorResponse, writeErr)
-		}
-	} else {
-		h.writeDatabaseError(w, err)
-	}
-}
-
-// writePropertyResponse writes a property response with OData context
-func (h *EntityHandler) writePropertyResponse(w http.ResponseWriter, r *http.Request, entityKey string, prop *metadata.PropertyMetadata, fieldValue reflect.Value) {
-	// Get metadata level to determine which fields to include
-	metadataLevel := response.GetODataMetadataLevel(r)
-
-	odataResponse := map[string]interface{}{
-		"value": fieldValue.Interface(),
-	}
-
-	// Only include @odata.context for minimal and full metadata (not for none)
-	if metadataLevel != "none" {
-		contextURL := fmt.Sprintf("%s/$metadata#%s(%s)/%s", response.BuildBaseURL(r), h.metadata.EntitySetName, entityKey, prop.JsonName)
-		odataResponse[ODataContextProperty] = contextURL
-	}
-
-	// Set Content-Type with dynamic metadata level
-	w.Header().Set(HeaderContentType, fmt.Sprintf("application/json;odata.metadata=%s", metadataLevel))
-	w.WriteHeader(http.StatusOK)
-
-	// For HEAD requests, don't write the body
-	if r.Method == http.MethodHead {
-		return
-	}
-
-	if err := json.NewEncoder(w).Encode(odataResponse); err != nil {
-		fmt.Printf("Error writing property response: %v\n", err)
-	}
-}
-
-// writeRawPropertyValue writes a property value in raw format for /$value requests
-func (h *EntityHandler) writeRawPropertyValue(w http.ResponseWriter, r *http.Request, prop *metadata.PropertyMetadata, fieldValue reflect.Value) {
-	// Set appropriate content type based on the value type
-	valueInterface := fieldValue.Interface()
-
-	// Check for binary data ([]byte) first
-	if fieldValue.Kind() == reflect.Slice && fieldValue.Type().Elem().Kind() == reflect.Uint8 {
-		// Binary data - set appropriate content type and write raw bytes
-		// Use custom content type if specified, otherwise default to application/octet-stream
-		contentType := "application/octet-stream"
-		if prop.ContentType != "" {
-			contentType = prop.ContentType
-		}
-		w.Header().Set(HeaderContentType, contentType)
-		w.WriteHeader(http.StatusOK)
-
-		// For HEAD requests, don't write the body
-		if r.Method == http.MethodHead {
-			return
-		}
-
-		// Write raw binary data
-		if byteData, ok := valueInterface.([]byte); ok {
-			if _, err := w.Write(byteData); err != nil {
-				fmt.Printf("Error writing binary value: %v\n", err)
-			}
-		}
-		return
-	}
-
-	// Determine content type based on the property type
-	switch fieldValue.Kind() {
-	case reflect.String:
-		w.Header().Set(HeaderContentType, ContentTypePlainText)
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64,
-		reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64,
-		reflect.Float32, reflect.Float64:
-		w.Header().Set(HeaderContentType, ContentTypePlainText)
-	case reflect.Bool:
-		w.Header().Set(HeaderContentType, ContentTypePlainText)
-	default:
-		// For other types, use application/octet-stream
-		w.Header().Set(HeaderContentType, "application/octet-stream")
-	}
-
-	w.WriteHeader(http.StatusOK)
-
-	// For HEAD requests, don't write the body
-	if r.Method == http.MethodHead {
-		return
-	}
-
-	// Write the raw value
-	if _, err := fmt.Fprintf(w, "%v", valueInterface); err != nil {
-		fmt.Printf("Error writing raw value: %v\n", err)
-	}
-}
-
-// applyStructuralPropertySelect applies SELECT clause to fetch only the structural property and key columns
-func (h *EntityHandler) applyStructuralPropertySelect(db *gorm.DB, prop *metadata.PropertyMetadata) *gorm.DB {
-	// Build select columns list: property + all key properties
-	// Use struct field names - GORM will handle column name conversion
-	selectColumns := []string{prop.Name}
-	for _, keyProp := range h.metadata.KeyProperties {
-		// Avoid duplicates if the property itself is a key
-		if keyProp.Name != prop.Name {
-			selectColumns = append(selectColumns, keyProp.Name)
-		}
-	}
-	return db.Select(selectColumns)
-}
-
 // fetchParentEntityWithNav fetches the parent entity and preloads the specified navigation property
 func (h *EntityHandler) fetchParentEntityWithNav(entityKey, navPropertyName string) (interface{}, error) {
 	parent := reflect.New(h.metadata.EntityType).Interface()
 
 	var db *gorm.DB
 	var err error
-	
+
 	// Handle singleton case where entityKey is empty
 	if h.metadata.IsSingleton && entityKey == "" {
 		// For singletons, we don't use a key query, just fetch the first (and only) record
@@ -1387,7 +689,7 @@ func (h *EntityHandler) writeNavigationCollectionRefFromData(w http.ResponseWrit
 func (h *EntityHandler) handlePutNavigationPropertyRef(w http.ResponseWriter, r *http.Request, entityKey string, navigationProperty string) {
 	// Parse the navigation property to extract any key (though PUT shouldn't have a key in the navigation property)
 	navPropName, targetKey := h.parseNavigationPropertyWithKey(navigationProperty)
-	
+
 	// If a key was provided in the navigation property name for PUT, that's an error
 	if targetKey != "" {
 		if err := response.WriteError(w, http.StatusBadRequest, "Invalid request",
@@ -1472,7 +774,7 @@ func (h *EntityHandler) handlePutNavigationPropertyRef(w http.ResponseWriter, r 
 func (h *EntityHandler) handlePostNavigationPropertyRef(w http.ResponseWriter, r *http.Request, entityKey string, navigationProperty string) {
 	// Parse the navigation property to extract any key (though POST shouldn't have a key in the navigation property)
 	navPropName, targetKey := h.parseNavigationPropertyWithKey(navigationProperty)
-	
+
 	// If a key was provided in the navigation property name for POST, that's an error
 	if targetKey != "" {
 		if err := response.WriteError(w, http.StatusBadRequest, "Invalid request",
